Fixes  #138
The problem was a python namespace lifetime issue.
inside do_config  there was a line:
config = SimpleNamespace()

this was inside an if statement.  Being in an if statement it was not replacing the config that was passed in,  it was creating a new one that shadowed the old version in a new variable context.
We had two choices to fix this:
1) instead of creating a new simpleNamespace, we could have emptied the old one.
2) return the new simpleNamespace from within the context of the if statement.

i chose 2 as it highlights the namespace issues.